requirements.txt
# dev env requirements
# build
pyyaml
# docs
Sphinx==3.0.2
sphinx-automodapi==0.12
nbsphinx==0.6.1
sphinx-rtd-theme==0.4.3 # <- not on conda repo so conda install will fail; fine for pip though
# test
testfixtures
# examples
ipywidgets
plotly
